Add new Yocto layer for common packages to IVI systems
'meta-ivi-common' is a new Yocto layer for common packages
to various Linux-based In-Vehicle Infotainment system.

The metadata to install common packages which developed by other OSS
projects (except Poky/OE) would be cloned here.  (This cloning is
tentative until a better way to refer those metadata is found.)

To add common packages which are already in Poky/OE into rootfs
of meta-agl, add them to packagegroup-ivi-common-* in meta-ivi-common.

And the common packages which used to be in packagegroup-agl-core-*
are moved to packagegroup-ivi-common-*. (weston and weston-examples)

Change-Id: I94ea7117029cc7769583358e1e2ec0aad2d21e5e
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>