Copy metadata for GStreamer1.0 from Yocto GENIVI Baseline
There are 2 type of metadata for GStreamer when building
GDP/GENIVI baseline, meta-ivi (7.0).

One is metadata in meta-ivi which is to install ver.1.2.3 ,
the other is in poky/meta is to install ver.1.4.1.

By default, GENIVI baseline preferred 1.4.x by macro
PREFERRED_VERSION_ in poky-ivi-systemd.conf.

As a result, for QEMU, GStreamer1.0_1.4.x is used and it works.

On the other hand, Renesas BSP (proprietary multimedia packages)
requires GStreamer1.0_1.2.x and it override PREFERRED_VERSION_
then select to use 1.2.x instead of 1.4.x.

From the above, AGL Distro also need metadata for GStreamer1.0_1.2.x
to run software on R-Car Gen2 Porter board.

This is the first step to support GStreamer in AGL Distro.

All files in this patch are from:
    URL: git://git.yoctoproject.org/meta-ivi
    commit: 0d780d0cfd38694ae5e6f0198adcb72684b01acc

NOTE: This patch is not completed alone and some error will be
got when building with meta-renesas. Please wait a subsequent patches.

Change-Id: I8ce1ec2b1593fc59780495459afc8c642e2d80d6
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>