Add a minimal set of packages for GStreamer1.0
This configuration is for installing a minimal set of packages
to use GStreamer1.0 in meta-agl.

This configuration also switch 2 versions of GStreamer,
when building for QEMU, GStreamer1.0 1.4.1 is selected by AGL distro
configuration, for R-Car2 Porter, GStreamer1.0 1.2.3 by meta-renesas.

The installed plugins of GStreamer1.0 is also minimal, so when more
are needed, add appropriate packages to IMAGE_INSTALL in your
local.conf manually.

Change-Id: I379e75f2eaacd47e80ceafdb00e2dcfaeda0db1f
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>