Provide Wayland-IVI-Extension 1.5.0
Wayland-IVI-Extension provides an additional shell plugin
for IVI-Shell, along with command-line tools and
demonstration applications.

Project site :
http://projects.genivi.org/wayland-ivi-extension

We make this recipe a member of the "ivi" package group,
which will always be built.
We do not activate it by default though, to facilitate
development and debugging at this point ; but add a comment
in "weston.ini" allowing to enable it easily.

Change-Id: Id1c788880d730ed612a55c19d320e61e170c792c
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>