Adds meta-agl-security layer
This commits adds the security layer within the sources
of the AGL project.

This layer, named "meta-agl-security", is intended to host
receipes specific to the security framework of AGL.

In the case where a choice exist between several security
framework, the switch is done using "distro features" in
configuration files.

This layer will either hold the several framework and their
specific common components.

Change-Id: I8a34b2d6c06cd7d2687825a54d1fe75441de2895
Signed-off-by: José Bollo <jose.bollo@iot.bzh>