Provide Wayland FITS, disabled by default
Wayland FITS is a functional integration test suite for
Wayland/Weston.
It consists in 2 components : a Weston plugin, and a
"wfits" binary used to run tests. The tests themselves
simulate user interaction with surfaces and windows, using
virtual keyboard/mouse/touchscreen drivers.

The recipe was recently submitted to the official
"meta-openembedded" layer, and can be found here :
http://lists.openembedded.org/pipermail/openembedded-devel/
2015-September/103511.html

The package will not be built unless the user explicitly
adds "wayland-fits" to his "conf/local.conf" file.

QEMU needs the "uinput" driver available on startup, so we
always build it, but only install it if the package has been
selected.

Change-Id: Ib6045007f6c7e1e22b48607a66930ce341c58002
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>