Fix Weston IVI-Shell layout on Renesas R-Car E2 Silk
On Renesas Silk boards, when running weston ivi-shell,
layout is too small and black borders are visible.
This is due to the presence of a VGA adapter, which
ivi-shell wrongly queries for resolution size values.

Work around this ; this patch is not upstreamed due to
current work being done on ivi-shell multiscreen support
(https://github.com/ntanibata/weston-ivi-shell/commits/
multiscreen_feature_03).

PS : although this patch is primarily intended for Silk,
we put it into "meta-agl" because it is a general-purpose
fix. It would also fix the issue on laptops having VGA
adapters, e.g..

Change-Id: I25447b54d53eb8f8b7261903c0699a53c81790f8
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>