Fix compilation of AMB if icecc is installed on the build machine.
If the buildhost has icecc installed, AMBs cmake file will
hardcode the compiler to icecc (understandable, but not a good
idea for crosscompiling). In our case this leads to
the replacement of the compiler supplied by the build environment
with icecc. This breaks the compilation.

We turn off the icecc detection, which is unfortunately on by default.

Plus fixes for a few typos.

Change-Id: Ibc28d3d697f1769f134e2a8f630e42058255bf3b
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>