AMB (automotive-message-broker): enable gpsnmea plugin
In this version of AMB, GPS plugin is named "gpsnmea" and
does not really depend on gpsd, so fix the build option and
remove the dependency.

If gpsnmea is enabled (default), we should ship a matching
configuration file so that it gets loaded automatically.
Consider an USB GPS device ; at worst, if no device is
present, harmless NULL values will be returned by AMB.

Change-Id: I60fc960bff04b334a2556b9ec92a4089398d453e
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>