Add touch_frame events to Weston 1.5.0
Otherwise no Qt applications react on touch. This issue was
eventually resolved upstream by dropping the built-in evdev backend
and relying on libinput exclusively, where it is fixed. The libinput
backend is required since Weston 1.7. In 1.6 the libinput backend
was made the default, but the old backend was still there. Thus this
fix also applies to Weston 1.6.

Signed-off-by: Risto Avila <risto.avila@theqtcompany.com>