Remove Yocto 1.7 compatibility crufts
As we are now based upon Yocto 2.0, we can remove lots of
backported recipes and contextual patches, which only made
sense in the context of Yocto 1.7.

Here is the rationale behind each change:

- gupnp-dlna: already in meta-openembedded/jethro
- ofono: BlueZ 5 compatibility is now working
- rygel: already in meta-openembedded/jethro
- libgee: already in meta-openembedded/jethro
- pulseaudio: BlueZ 5 compatibility is now working
- lightmediascanner: already in meta-openembedded/jethro
- gpsd: BlueZ 5 compatibility is now working
- libmediaart: already in meta-openembedded/jethro
- gtest: already in meta-openembedded/jethro
- wayland-fits: already in meta-openembedded/jethro

PS: "meta-openembedded/meta-efl" is now required because we
require the "lightmediascanner" recipe, which is unrelated
to EFL but happens to live there.

Change-Id: Ib46d152f80b1a8ee5cf848a41e408ece28e6ec65
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>