Add IVI-specific PulseAudio patches and development files
This is a stripped-down port of Tizen IVI PulseAudio 5.0
patches to AGL PulseAudio 6.0.
These patches mainly:
- provide a new "pulseaudio-module-dev" package, including
headers and libraries for out-of-tree module development;
- add generic helper functions;
- add a helper "libvolume" library;
- add helper functions for an efficient "volume ramp
up/down" feature, so that a module does not need to track
volume changes' timing and state in realtime.

These patches are needed for the PulseAudio Routing Module
described on: https://wiki.automotivelinux.org/
eg-ui-graphics-req-audiorouting (which is itself a stripped
-down version of the Tizen IVI Audio Routing module).

Change-Id: Ib826f3106b91ffdef639c2719a580373f84449c7
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>