Add agl-audio-plugin (PulseAudio Routing plugin)
agl-audio-plugin is a stripped-down version of the former
Tizen IVI Audio Routing plugin (also known as "module-
murphy-ivi"). This is stripped down is the sense that:
- it does not require Murphy anymore;
- it does not implement the most complex use cases anymore;
- it is able to work standalone by using a JSON
configuration file or its own embedded configuration.

It still has the basic structure needed to interconnect an
external policy engine (such as GENIVI AudioManager).

Change-Id: I133916cada5cc928187f2aef6448fdaf7c03c3c6
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>