Add PulseAudio server and systemd service
This commit adds the PulseAudio server to the image, and a
related systemd service for automatic startup.

This change is motivated by an Audio Routing model
involving a PulseAudio plugin and a AudioManager interface,
as described on: https://wiki.automotivelinux.org/
eg-ui-graphics-req-audiorouting .

PulseAudio server is launched in user mode.
("paplay <file>.wav" may verify the functionality)

Change-Id: I12e2bfb2af5494e530e29bda77b0d88999cf7b4b
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>