Fix issue that pulseaudio freezes when pausing audio
This patch is ported from Debian/Ubuntu.

When using glibc-2.21 or later on linux kernel < 3.14.3,
this issue should be reproduced.

Original patch is here:
https://sources.debian.net/patches/glibc/2.22-5/arm/local-arm-futex.diff/

AGL needs this patch because the version of kernel for
porter board is 3.10 (later kernel is not supported).

This patch and recipe is a provisional until upstream
(poky) would be fixed.

Thanks Isogai-san to find and test this patch.

v2: changed bbappend to use SRC_URI_append_arm
    and used glibc_2.2%.bbappend as wildcard for glibc version
    (jsmoeller)

Change-Id: I4a2a0ad8fe3927000b6826e0b1bcd214cf8c92b6
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>