aglsetup.sh: handle machines/features from any layer
This patch allows the setup script called by scripts/aglsetup.sh to
search for available machines and features in all repositories located
in the meta directory (the directory where meta-agl has been cloned).

In particular, this now allows to specify a machine or a feature located
in meta-agl-devel.

Important note: duplicate machines or features in different layers
are detected and forbidden (script fails with error) to avoid any ambiguity.

Change-Id: If55105a05889ca95da856d65986a86f4ef5f4b2c
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>