meta-agl-bsp/meta-renesas: weston.ini for Reneses
Add weston.ini with preconfigured but commented
out screen resolutions in the output section.
Weston 1.9 with DRM backend on Renesas Porter
board is known to be picky about monitors and
screen resolutions. This way the developer can
quickly switch to another configuration if the
default fails on the current monitor. This patch
should be considered as a temporary work around
until there is a generic solution.

Bug-AGL: SPEC-298

Change-Id: I772945a083ef4cab240c49e03c5f0cf1172a6ab2
Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>