dra7xx-evm: u-boot: Add custom uenv.txt
dra7xx-evm supports multiple LCD panels, to boot the board
successfully, the uenv.txt has to specify the correct dtb
file. This patch adds the default dtb and specifies one of
the LCD panels in it.

NOTE:
It could happen that the panel that you have is a different
panel, in that case you would have to manually edit the same.

Change-Id: I0fdbe0ce48ca0654604a5808de684558308582c5
Signed-off-by: Karthik Ramanan <a0393906@ti.com>