Set rootfs size based on real size of rootfs image
ROOTFS_SIZE is recomputed for each image type based on total size of
rootfs directory. If rootfs directory gets smaller between creation of
rootfs image and SD image, it fill result in corrupted SD image.
It should not happen in theory, but it happens occasionally for a reason
I could not identify.

Change-Id: I1c888c179fdcec32c3cacbbce7b2d34ec70bbb17
Signed-off-by: Anton Gerasimov <anton@advancedtelematic.com>