dra7xx-evm: u-boot: Change default supported display to OSD
The latest dra7xx-evm purchased from ti.com comes with the OSD
display, this change will allow the AGL nightly images to work
on the latest dra7xx-evm by default.

If the user has an older EVM, then the uEnv.txt should be modified
to match the connected displays. The complete list of supported
LCDs can be found by checking the /boot folder of the the rootfs.

Change-Id: I7f7ac06a49670eb8b7446a6b27854bee4b76ace6
Signed-off-by: Karthik Ramanan <a0393906@ti.com>