meta-ti: weston: remove unneeded ivi shell patches and use standard ivi-controller
The meta-arago layer included a number of ivi shell patches that are
not needed in AGL. One multi screen patch conflicts with another
bbappend causing build issues as it's already included by a less
specific layer. Further, the added "simple TI layout controller",
autolaunch, and screenshooter features are all unneeded in AGL. These
are removed to reduce the amount of non-essential patches we carry within
AGL. As a result of this, all weston.ini directives which use the
"simple TI layout controller" directives are removed and weston.ini
is switched to use the standard ivi-controller so that WindowManager may
run and access the ILM API. The end result is that WindowManager and
HomeScreen may now be launched after boot successfully.

Change-Id: I8d777fbd4c02acbf9d3ccaeb215b651a8562d5bb
Signed-off-by: Matt Porter <mporter@konsulko.com>