Test - mitigation of tar-native issue
This changeset removes tar-native from ASSUME_PROVIDED in an attempt
to mitigate the tar-native extraction issue that we see on the jenkins slaves.

Is I understand the race-window, tar-native is just being extracted and already
in a preferred PATH while being extracted. The right fix would likely be
to extract first, then change PATH after the file is in place. A way to
do this is being researched.

Change-Id: I0b138a1bd27406e719ab0d711cf542a812fb8460
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>