Disable Weston DRM backend in QEMU
Since the Weston DRM backend does not work with the video card implementations
in VirtualBox and VMWare player, and has some outstanding issues on QEMU that
have resulted in GenIVI not enabling it[1], disable it for now. A comment has
been added to the QEMU weston.ini indicating that the DRM backend can be
re-enabled to potentially gain higher performance.

Additionally, the common weston.service systemd unit has been modified to
add the --tty option to weston-launch to start weston on /dev/tty1. This is
now required to avoid console messages writing on top of the HomeScreen in
QEMU, and is harmless to operation on the other platforms.

[1]: https://github.com/GENIVI/meta-genivi-dev/pull/34

Bug-AGL: SPEC-405

Change-Id: Ic0c0adeb15baf392e3ffa0bbd26c62d060df1ef5
Signed-off-by: Scott Murray <scott.murray@konsulko.com>