Splice out distro configuration in own layer for compatibility
Upstream is working on guidelines for DISTRO and BSP layers.
One outcome is the yocto-compat-layer tool.

A requirement for BSP and DISTRO layers is that they are single-purpose
aka just defining the distro and only defining the BSP.

To comply with this, move the distro out into meta-agl-distro and adapt the templates.

Bug-AGL: SPEC-472

Change-Id: I921127edc78e677bd75d42a793f0b3708a54dd79
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>