Ensure that eXtended Attributes are managed
Common tools like 'useradd' work better in a Smacked
context when the extended attributes are correctly
managed. This is achieved by enforcing extended
attributes "xattr" to be in DISTRO_FEATURES.

When the commit 1c3eae5e654b7942b3ffd0b53426ca77219cec03
of poky will be integrated in the upstream of AGL (pyro),
it will be possible to removes the 3 last lines of
meta-app-framework/conf/include/agl-appfw-smack.inc
that explicitely activate the handling of extended
attributes for the package "shadow". See SPEC-475.

Change-Id: I299e68a34e7e49b33f1046aa99d255655fa2ffcc
Signed-off-by: José Bollo <jose.bollo@iot.bzh>
Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>