Fix virtual provider for webruntime
A recipe that is to be used as virtual provider needs to declare
this in a PROVIDES += 'virtual/foo' statement.

The selection then happens through a PREFERRED_PROVIDER.

Add the missing PROVIDES.

Additionally we found that in case of web-runtime,
_R_PROVIDES is not reflecting the PROVIDES as expected 
and added the necessary override there, too.

Kudos to co-investigator Stéphane Desneux.

Change-Id: Ic9224ea122c21dab6ab2e7e515d6ac4d0417ef57
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>