Make /usr/local compatible with SOTA
This provides a link from /usr/local to ../../var/local
that enforces file hierarchy local to the machine
to be in a directory frankly separated from /usr.

This is important for SOTA/OSTree that must not manage
the locally installed files. But this is also important
for correctly separating layers of the target. For
this reason, the change is not conditionnal to SOTA.

Bug-AGL: SPEC-359
Bug-AGL: SPEC-533

Change-Id: I0a709ba15582a011a43f3a3b68d4230bae11b658
Signed-off-by: José Bollo <jose.bollo@iot.bzh>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/9071
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Stéphane Desneux <stephane.desneux@iot.bzh>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>