Add machine templates for the beaglebone
This adds the machine templates for the beaglebone board.
The beaglebone is a YP reference system with very good
support out-of-the box. It might be too slow for the
graphical demos, but could serve well in headless setups.

v2: remove whitespace.

Bug-AGL: SPEC-590
Change-Id: I3f1d4f57d2ce01f2603446b6663fdaa636a1f7be
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/9337
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Karthik Ramanan <a0393906@ti.com>
Reviewed-by: Stéphane Desneux <stephane.desneux@iot.bzh>
Reviewed-by: Martin Kelly <mkelly@xevo.com>