Allows specific features for agl-devel
Having features allowed only during developement
of AGL is a plus. This commit allows to add tests
within recipes to check if the current build is
AGL developement. If not the build is for production.
By default, extra feature must be off.

This commits doesn't add

  CPPFLAGS += "-DAGL_DEVEL"

to improve build time and cache space of yocto.

However, it is a good practice, when needed, to
use the tag AGL_DEVEL in C/C++ codes specific to
AGL and thus to include the line

  CPPFLAGS_append_agl-devel = " -DAGL_DEVEL"

in recipes specific to AGL.

AGL-Bug: SPEC-594

Change-Id: Ie967a4317886ee78304999f36de5db9de22568a7
Signed-off-by: José Bollo <jose.bollo@iot.bzh>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/9393
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Stéphane Desneux <stephane.desneux@iot.bzh>
Reviewed-by: Changhyeok Bae <changhyeok.bae@lge.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>