run-agl-postinsts: run before systemd-user-sessions
Add setting of Before to the template scheme in the run-agl-postinsts
recipe, and use it to make sure run-agl-postinsts runs before
systemd-user-sessions.  This ensures that on first boot all of the
applications will be installed before HomeScreen is started in the
user session.

Bug-AGL: SPEC-580

Change-Id: I08026f225971e1fe38e984dfed5fc2656955942c
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/9439
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>