pulseaudio: fix service/socket unit installation
It's not obvious how the base pulseaudio package was being pulled
in before to pick up the pulseaudio.{service,socket} symlinks in
/home/root/.config, but it's definitely not being installed now,
breaking PulseAudio usage in all applications.  Move the symlinks to
the pulseaudio-server package which is definitely always installed,
and also contains the target files for the symlinks.  Additionally,
move the symlinks to /usr/lib/systemd/user to prepare for the eventual
running of all applications as non-root and be compliant with
SPEC-505.

Bug-AGL: SPEC-612

Change-Id: Ic8e283e6ee32e3b532e93344445b239ed8d3e896
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/9497
Reviewed-by: Matt Porter <mporter@konsulko.com>
Reviewed-by: Matt Ranostay <matt.ranostay@konsulko.com>
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>