Fix i.MX6
Bug-AGL: SPEC-704

This is part of the fixe to make i.MX6 buildable again.
Following fixes are included in this change:

- adding build target imx6qsabreauto
- fixing qt5.8 build error
- refine bblayers.conf layer order (for dynamic-layers)
- meta-fsl-arm -> meta-freescale layer name change
- weston.ini lines (use fbdev-backend.so)
- /dev/galcore permission (by /etc/udev/rules.d/zz-dri.rules)
- kernel configuraion to include smack, etc.

Follwings are stil not addressed yet:

- bblayer.conf layer order still have issue re. dynamic-layers
- built and booted on imx6qsabreauto only yet.
  following existing community boards may need update
  (i have imx6qsabrelite, but no notrogen6x)

Note, this fix is prepared independently of nxp or fsl-community,
who should follow-up if there are mistakes to be fixed, or updates
that they find needed.

Update:

Refined on review comment regarding tune settings and
kernel configurations.  The board successfully boots and shows
AGL homescreen.

But I found I still have problem with USB.
lsusb says 'unable to initialize libusb: -99' error which I am
still faling to find the cause.

Update2:

usb issue may be related to dtb configuration in linux-fslc-imx
better to be tracked separately

Change-Id: If5fa4e5f0450385bcb2648d57e63dc1fc14b0574
Signed-off-by: Takashi Matsuzawa <tmatsuzawa@xevo.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/9971
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>