Executing /bin/bash explicitly
To resolve issue of SPEC-699, afm-install began to use
PIPESTATUS[] to catch error from dbus-send.

But the /bin/sh is still specified even though it is an unique
function of Bash.

Note: Because current AGL has Bash and /bin/sh is symbolic
link to /bin/bash, it works with no error before this fix applied.

Bug-AGL: SPEC-699

Change-Id: I730de85d593f6371210cf1424e7bc39ca9365f2f
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/10055
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>