backport open-vm-tools recipe
Without open-vm-tools, VMWare guests are very slow. open-vm-tools was
recently added to meta-openembedded, so we should get it when Rocko is
released. Until then, this patch backports it.

Note that open-vm-tools has an optional dependency on fuse, so we need
to include the meta-filesystems layer too.

Bug-AGL: SPEC-739

Change-Id: Icc5bf7f10b8e21778ab7043dea8d340eb41a527c
Signed-off-by: Martin Kelly <mkelly@xevo.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/10211
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>