agl-devel: auto-add drivers for vmdk images
Currently, if you build a vmdk file, you also need to manually add in
drivers to make your image fast (vbox/vmware drivers). We can make this
automatic so that when you build a vmdk, drivers are automatically added
in. Using this approach, both vbox and vmware drivers are added, which
is more than is generally necessary, but it means that a single vmdk
image can support many use cases, and the same image can be distributed
to many developers.

For developer images, this seems like a good tradeoff because
convenience is more important than disk space/bloat for developers. For
production, we leave this option off to reduce unneeded components.

Change-Id: I8139a3ed94599c1732ea8b58bbc735c0de594234
Signed-off-by: Martin Kelly <mkelly@xevo.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/10219
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>