raspberrypi: Fix U-Boot support
In order to use U-Boot on the raspberrypi3 we need to have the equivalent
of the following commit from upstream meta-raspberrypi:

commit 67e02164929a7d8cc187315aaaab60f6f3aaf422
Author: Paul Barker <paul@paulbarker.me.uk>
Date:   Mon Feb 13 23:32:43 2017 +0000

    u-boot: Fix boot on rpi3

With that done, we can re-enable U-Boot support for raspberrypi3.  With
this done we can also now re-enable testing, so do so. For consistency we
can now switch raspberrypi2 over to U-Boot as well.

Change-Id: I9c96a66bf5644bf63f2566559ab7d9c4101b7a55
Bug-AGL: SPEC-409
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/10491
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Kevin Hilman <khilman@baylibre.com>
ci-image-boot-test: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>