af-main: Fix 2 bugs
Allow packaging of companion programs
-------------------------------------

Allow to set executable permission on
files of the widget. Then the files can be run by
using their relative path within the widget.

Bug-AGL: SPEC-384
Bug-AGL: SPEC-785

Allows content in subdirectory "htdocs"
---------------------------------------

Fix the bug SPEC-808. The content of pure or hybrid HTML
widget had to be in the subdirectory "htdocs" but the attribute
"src" of the tag <content> shouldn't include the prefix "htdocs/".
It caused installation to fail because it was checking that the
file existed but in the root directory.

This commit fixes the paradox by also trying to search the
content in the subdirectory "htdocs". This is a temporary fix.

Bug-AGL: SPEC-808

Change-Id: I512e47d35c135329ef99e3617909ba6fe78eaca7
Signed-off-by: José Bollo <jose.bollo@iot.bzh>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/10603
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>