Do not use cpio.gz for initramfs as smack does not work
This changeset removes the wrong hardcoded INITRAMFS_FSTYPES on 3 boards.
It was moved over from the previous defaults prior to the change
to the AGL-wide default FSTYPES.

The correct image to use is ext4.gz (which is a default in poky-agl.conf) -
thus we can just remove these in the board includes.

Therefore the intel targets need to define INITRD_LIVE.

Bug-AGL: SPEC-836

Change-Id: If45f67d2bbd02331f1aea848e3230ce1e190c50d
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/10711
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Dominig ar Foll <dominig.arfoll@fridu.net>