meta-agl-bsp: meta-ti: Rework Weston patches for 2.0
We can drop the changes for the compositor to support RGB565 as this is now
upstream.  The other compositor patch is no longer relevant.  The soc
performance patch needs additional work to link but is also not required for
overall functionality so leave that broken.  For the rest of the patches we
just need to make minor changes about the location of the source files now as
the code has reorganized since the patches were written.

Bug-AGL: SPEC-908

Change-Id: I7034219510821b8f720328318e3e560783cf16d6
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11019
Reviewed-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-by: Matt Ranostay <matt.ranostay@konsulko.com>
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>