meta-intel: Include and deploy a startup.nsh file
On EFI systems we need to include a startup.nsh file that EFI will load
and run, in the case of not having the device we're booting from be
defined in the boot order.  We have the kernel provide this as it is
independent of boot loader.

Bug-AGL: SPEC-898
Change-Id: Ifdf615e78012e2dbcbd87ee97e4c9d9a19ea01fc
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11101
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>