linux-intel: Pull in common configuration fragments
Rework things so that the common kernel configuration fragment
additions are in a new include file, linux-agl.inc, and both the
existing linux-yocto_%.bbappend and the new linux-intel_%.bbappend
in meta-intel pull it in to pick up them up.  This avoids hitting
other linux-* recipes, and allows the possibility of using the
include file in kernel recipe bbappends for other BSPs.

Change-Id: I8f4f67c444c47b6a978d95fab660d478d5f764c1
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11129
Reviewed-by: Thomas Rini <trini@konsulko.com>
Reviewed-by: Dominig ar Foll <dominig.arfoll@fridu.net>
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>