systemd: backport serialization change from v234
Bisecting between systemd v232 and v234 to find the source of
the improvement in behavior of the serial tty lead to commit
e266c068b5597e18b2299f9c9d3ee6cf04198c41.  Backport the change
to hopefully fix the failures.  This backport can be removed on
the uprev to rocko.

Bug-AGL: SPEC-544

Change-Id: I5e3bbd21447bd8a0df98ce2d2ff26c74c4bf1048
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11135
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>