dra7xx-evm: fix build by including linux-dtb.inc
meta-ti does not maintain a pyro branch. As such we are
relying on building with the master branch. This broke
due to commit 391bfb8c3002bbf5ba6a51d689a7f2ba511c6682
which removes support for linux-dtb.inc from the
linux-ti-staging recipe due to an oe-core change on
master. Add linux-dtb.inc to the bbappend to fix.

Bug-AGL: SPEC-932

Change-Id: I3fb6799a5fe4a083159b600a81764d67e5b22771
Signed-off-by: Matt Porter <mporter@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11149
Reviewed-by: Thomas Rini <trini@konsulko.com>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>