weston-ini-conf: Make this track variables correctly for signatures
bitbake does not know to track values that are found and used via
getVarFlags() without them being otherwise refernced and tracked in ways
that are otherwise caught.  Further given that we use multiple levels of
varflags here we need to be explicit in their tracking.  Add all
variables that we use to varflags where they are added to the metadata.

Bug-AGL: SPEC-939
Change-Id: I32cf16c4354d733e086c6ea9a845934beff94715
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11255
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>