weston-ini-conf: Fix task signatures, mark as machine-specific package
We need to do two things here.  First, bitbake does not know to track
values that are found and used via getVarFlags() without them being
otherwise refernced and tracked in ways that are otherwise caught.
Further given that we use multiple levels of varflags here we need to be
explicit in their tracking.  Add all variables that we use to varflags
where they are added to the metadata.  This change comes directly from
master.  The second thing is that we need to tell bitbake/OpenEmbedded
that this is a machine-specific package not architecture-generic.

Bug-AGL: SPEC-965
Change-Id: I748dbf40f37aaf7f4a6dd6ddd65257fdfcf83098
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11355
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-boot-test: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>