R-Car Gen3: Add WIC support
Add a new wks file for the R-Car Gen3 platforms.  Given that U-Boot is
on-device and not updated by us, we just need to make a single partition
that contains the rootfs.  This  matches the existing instructions.  We
need to include kernel-devicetree into the resulting image so that we
have the device tree there to be loaded in.

Bug-AGL: SPEC-899
Change-Id: Iad55ef0ed49e7a53846fa36393e71d16cd5e423d
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11391
Reviewed-by: Stéphane Desneux <stephane.desneux@iot.bzh>
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-boot-test: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Reviewed-by: Scott Murray <scott.murray@konsulko.com>