python-wand: Add patch to ensure our built ImageMagick is used
It can be slightly tricky to build a copy of python-wand that will also
function.  A problem is that without additional logic (MAGICK_HOME), the
python code will not look outside of system paths for a copy of the
imagemagick library.  However, even with this path added to the search
list, the code will still try a 'naked' load of the library it wants.
This can result in a visible failure if you have a new enough
imagemagic-6 installed (such as 6.9.7) as the date string checking logic
will fail.  Work around these problems by setting MAGICK_HOME globally
to the correct location and patch the api code to look for the form of
imagemagick that we build first, rather than last, so that we are going
to first search for what we know we built.

Bug-AGL: SPEC-984
Backport: https://lists.linaro.org/pipermail/openembedded/2017-October/000102.html
Change-Id: I9f0e3e64ab4601845e4f336aa81ce7c3d01ff7b7
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11587
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-boot-test: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>