Fix FTBS on newer pyro versions due to duplicate userids
This change removes the agl-users package from the
packagegroup-agl-core-security.bbappend .
The reason is that for SOTA we have to use predefined userids
to stay consistent across builds.
Newer pyro releases are more picky about duplicate userids
from static files and those being created on-the-fly.

For now we remove the dynamic variant by removing
agl-users from the packagegroup.

Change-Id: I54034361eec344b8130754fadf8c48207321797a
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/11715
Reviewed-by: Thomas Rini <trini@konsulko.com>
Reviewed-by: Stéphane Desneux <stephane.desneux@iot.bzh>