meta-intel: Switch to using WIC images only
The hddimg format has a hard-stop at 4GB sized images.  We can trigger
this rather easily at this point under agl-demo-platform-crosssdk.  WIC
images do not have any such limitation, so we remove hddimg from the
list of built images.  The main impact of this is that whereas before we
would modify the default kernel command line via APPEND in the conf file
we would now modify it via the wks file we create for each platform and
that mkefi-agl.sh is now redundant.  These wks files are based on the
systemd one but also reflect the design of mkefi-agl.sh in that they
drop the swap partition.  Much of the APPEND commentary is moved to the
wks file and we remove some irrelevant comments as well.

Bug-AGL: SPEC-963
Change-Id: If667e1f946e4b9793b63f6d7f572d56d73e35767
Signed-off-by: Tom Rini <trini@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/12159
ci-image-build: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>
ci-image-boot-test: Jenkins Job builder account <agl-jobbuilder@automotivelinux.org>