bsp: add salvator-x H3 support
The Salvator-X uses the same SoC as the H3 but includes other
peripherals. As a result, it can reuse the same h3ulcb config but needs
a different device tree and kernel config. This is already supported in
meta-rcar-gen3 but needs AGL plumbing, so add it.

There is also an M3 version of the Salvator-X, but I don't have one and
cannot test it. So I've added support only for the H3 version here.

Change-Id: I64683f4f65dda3fee63e6243f0ba916e568ca245
Signed-off-by: Martin Kelly <mkelly@xevo.com>